01-10-25 / refactor: Renomeando variáveis do .env e reestruturando pastas e arquivos
