+++ .env.example
@@ -1,12 +1,22 @@
 
-SECRET= aef01fecdefd2c0eacb0761e978e39866778eb2aa7bf0483606a9a4c41de2025ca63d9d39b9d6561827f9612f7ddc04eda54d7732797d4b6f78a74d4abca93e0
-DB_HOST= 10.89.240.82 / db
-DB_USER= alunods / root
-DB_PASSWORD= senai@604 / root
-DB_NAME= stock
-SALT_ROUNDS= 12
+# Variável de segurança principal da aplicação (JWT)
+SECRETKEY= aef01fecdefd2c0eacb0761e978e398e39866778eb2aa7bf0483606a9a4c41de2025ca63d9d39b9d6561827f9612f7ddc04eda54d7732797d4b6f78a74d4abca93e0
 
+# Configuração de custo para hashing de senhas com bcrypt
+BCRYPTSALTROUNDS= 12
+
+# Variáveis de Conexão com Banco de Dados
+DATABASEHOST= 10.89.240.82 / db
+DATABASEUSER= alunods / root
+DATABASEPASSWORD= senai@604 / root
+DATABASENAME= stock
+
+# Chave de API para serviços do Google Gemini
 GEMINIAPIKEY= AIzaSyC-cpyktB9SZtEMAitDKhbK82QPUpd0IN8
 
-GMAIL_USER= senaiestoque@gmail.com
-GMAIL_PASS= kqeq awbg zxxn lbjw
+# Credenciais de Conta de E-mail para envio (Gmail/SMTP)
+MAILUSERNAME= senaiestoque@gmail.com
+MAILPASSWORD= kqeq awbg zxxn lbjw
+
+# Configuração da Porta da API
+PORT= 5000
