+++ docker-compose.yml
@@ -5,7 +5,7 @@ services:
     restart: always
     environment:
       # Usando a variável do .env para o banco de dados
-      - MYSQL_ROOT_PASSWORD=${DB_PASSWORD}
+      - MYSQL_ROOT_PASSWORD=${DATABASEPASSWORD}
       - MYSQL_DATABASE=stock
     ports:
       - "3307:3306"
@@ -13,7 +13,7 @@ services:
       - db_data:/var/lib/mysql
       - ./sql/docker/init.sql:/docker-entrypoint-initdb.d/init.sql
     healthcheck:
-      test: ["CMD", "mysqladmin", "ping", "-h", "localhost", "-u", "root", "-p${DB_PASSWORD}"]
+      test: ["CMD", "mysqladmin", "ping", "-h", "localhost", "-u", "root", "-p${DATABASEPASSWORD}"]
       timeout: 20s
       retries: 10
       start_period: 30s
@@ -29,7 +29,7 @@ services:
       - /usr/app/node_modules
     environment:
       # Passando a variável do .env para a sua aplicação
-      - MYSQL_ROOT_PASSWORD=${DB_PASSWORD}
+      - MYSQL_ROOT_PASSWORD=${DATABASEPASSWORD}
       - MYSQL_DATABASE=stock
       - MYSQL_HOST=db
       - MYSQL_PORT=3306
