+++ src/db/connect.js
@@ -2,10 +2,10 @@ require('dotenv').config();
 const mysql = require("mysql2");
 
 const pool = mysql.createPool({
-  host: process.env.DB_HOST, 
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
+  host: process.env.DATABASEHOST, 
+  user: process.env.DATABASEUSER,
+  password: process.env.DATABASEPASSWORD,
+  database: process.env.DATABASENAME,
 });
 
 module.exports = pool;
